# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/check-policies/{id}" : {
      "get" : {
        "tags" : [ "check-policy-controller" ],
        "operationId" : "getById_1",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCheckPolicyDto"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "check-policy-controller" ],
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CheckPolicyUpdateCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCheckPolicyDto"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "check-policy-controller" ],
        "operationId" : "delete_1",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/v1/alert-rules/{id}" : {
      "get" : {
        "tags" : [ "alert-rule-controller" ],
        "operationId" : "getAlertRule",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAlertRuleDto"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "alert-rule-controller" ],
        "operationId" : "updateAlertRule",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AlertRuleUpdateCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAlertRuleDto"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "alert-rule-controller" ],
        "operationId" : "deleteAlertRule",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/contact-groups" : {
      "get" : {
        "tags" : [ "contact-group-controller" ],
        "operationId" : "getByWorkspace",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "isActive",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "createdAt"
          }
        }, {
          "name" : "sortDir",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "desc"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedResponseDtoContactGroupDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "contact-group-controller" ],
        "operationId" : "create",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContactGroupCreateCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseContactGroupDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/check-policies" : {
      "get" : {
        "tags" : [ "check-policy-controller" ],
        "operationId" : "getByWorkspace_1",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "expectedStatusCode",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "hasLatencyThreshold",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "createdAt"
          }
        }, {
          "name" : "sortDir",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "desc"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedResponseDtoCheckPolicyDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "check-policy-controller" ],
        "operationId" : "create_1",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CheckPolicyCreateCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCheckPolicyDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/alert-rules" : {
      "get" : {
        "tags" : [ "alert-rule-controller" ],
        "operationId" : "getAlertRules",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ruleType",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "operator",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "isActive",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "createdAt"
          }
        }, {
          "name" : "sortDir",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "desc"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedResponseDtoAlertRuleDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "alert-rule-controller" ],
        "operationId" : "createAlertRule",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AlertRuleCreateCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAlertRuleDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiResponseContactGroupDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ContactGroupDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ContactGroupDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "workspaceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "emailAddresses" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "CheckPolicyUpdateCommand" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "intervalSeconds" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 5
          },
          "timeoutMillis" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 100
          },
          "retryCount" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "failureThreshold" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "latencyThresholdMillis" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expectedStatusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expectedResponseBody" : {
            "type" : "string"
          },
          "responseRegex" : {
            "type" : "string"
          }
        },
        "required" : [ "failureThreshold", "id", "intervalSeconds", "name", "retryCount", "timeoutMillis" ]
      },
      "ApiResponseCheckPolicyDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/CheckPolicyDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "CheckPolicyDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "intervalSeconds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "timeoutMillis" : {
            "type" : "integer",
            "format" : "int32"
          },
          "retryCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "failureThreshold" : {
            "type" : "integer",
            "format" : "int32"
          },
          "latencyThresholdMillis" : {
            "type" : "integer",
            "format" : "int32"
          },
          "workspaceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "expectedStatusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expectedResponseBody" : {
            "type" : "string"
          },
          "responseRegex" : {
            "type" : "string"
          }
        }
      },
      "AlertRuleUpdateCommand" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "ruleType" : {
            "type" : "string",
            "enum" : [ "CONSECUTIVE_FAILURE", "RESPONSE_TIME_EXCEEDED", "STATUS_CODE_MISMATCH" ]
          },
          "operator" : {
            "type" : "string",
            "enum" : [ "GT", "LT", "EQ", "NE" ]
          },
          "thresholdValue" : {
            "type" : "number",
            "format" : "double"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "contactGroupIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "overrideDefaultContacts" : {
            "type" : "boolean"
          }
        },
        "required" : [ "id", "name", "operator", "ruleType", "thresholdValue" ]
      },
      "AlertRuleDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "ruleType" : {
            "type" : "string",
            "enum" : [ "CONSECUTIVE_FAILURE", "RESPONSE_TIME_EXCEEDED", "STATUS_CODE_MISMATCH" ]
          },
          "operator" : {
            "type" : "string",
            "enum" : [ "GT", "LT", "EQ", "NE" ]
          },
          "thresholdValue" : {
            "type" : "number",
            "format" : "double"
          },
          "workspaceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "contactGroupIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "overrideDefaultContacts" : {
            "type" : "boolean"
          }
        }
      },
      "ApiResponseAlertRuleDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AlertRuleDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ContactGroupCreateCommand" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "emailAddresses" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "name" ]
      },
      "CheckPolicyCreateCommand" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "intervalSeconds" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 5
          },
          "timeoutMillis" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 100
          },
          "retryCount" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "failureThreshold" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "latencyThresholdMillis" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expectedStatusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expectedResponseBody" : {
            "type" : "string"
          },
          "responseRegex" : {
            "type" : "string"
          }
        },
        "required" : [ "failureThreshold", "intervalSeconds", "name", "retryCount", "timeoutMillis" ]
      },
...
      "AlertRuleCreateCommand" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "ruleType" : {
            "type" : "string",
            "enum" : [ "CONSECUTIVE_FAILURE", "RESPONSE_TIME_EXCEEDED", "STATUS_CODE_MISMATCH" ]
          },
          "operator" : {
            "type" : "string",
            "enum" : [ "GT", "LT", "EQ", "NE" ]
          },
          "thresholdValue" : {
            "type" : "number",
            "format" : "double"
          },
          "contactGroupIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "overrideDefaultContacts" : {
            "type" : "boolean"
          }
        },
        "required" : [ "name", "operator", "ruleType", "thresholdValue" ]
      },
...
      "ApiResponsePagedResponseDtoContactGroupDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedResponseDtoContactGroupDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PagedResponseDtoContactGroupDto" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ContactGroupDto"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalItems" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ApiResponsePagedResponseDtoCheckPolicyDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedResponseDtoCheckPolicyDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ApiResponsePagedResponseDtoAlertRuleDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedResponseDtoAlertRuleDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}